Abstract the price total out of the product class
- Abstract the price total out of the product class
- Added ProductViewCell super class to cut down the need for repeat code.
- naming scheme for Double extension method "toCurrency()"
