+++ swift/ios-shoppe-demo/Model/Product.swift
@@ -12,8 +12,7 @@ import UIKit
 class Product {
     var title: String
     var description: String
-    var rawPrice: Double
-    var price: String
+    var price: Double
     var imageName: String
     var image: UIImage = UIImage()
     var unit: String
@@ -22,13 +21,12 @@ class Product {
     init(_ dict: [String: Any]) {
         self.title = dict["title"] as? String ?? ""
         self.description = dict["description"] as? String ?? ""
-        self.rawPrice = dict["price"] as? Double ?? -1.0
-        self.price = rawPrice.convertDoubleToCurrency()
+        self.price = dict["price"] as? Double ?? -1.0
         self.imageName = dict["image"] as? String ?? ""
         self.unit = dict["unit"] as? String ?? ""
     }
+}
 
-    func getTotal() -> Double {
-        return rawPrice.multiply(quantity)
-    }
+func getTotal(for product: Product) -> Double {
+    return product.price.multiply(product.quantity)
 }
