+++ swift/ios-shoppe-demo/Views/ProductSummaryCell.swift
@@ -10,22 +10,4 @@ import Foundation
 import UIKit
 
 
-class ProductSummaryCell: UITableViewCell {
-    @IBOutlet weak var titleLabel: UILabel!
-    @IBOutlet weak var quantityLabel: UILabel!
-    @IBOutlet weak var totalLabel: UILabel!
-
-    var product: Product! {
-        didSet {
-            setup()
-        }
-    }
-
-    func setup() {
-        DispatchQueue.main.async {
-            self.titleLabel.text = self.product.title
-            self.quantityLabel.text = "x\(self.product.quantity)"
-            self.totalLabel.text = self.product.getTotal().convertDoubleToCurrency()
-        }
-    }
-}
+class ProductSummaryCell: ProductViewCell {}
